Add board_info impl
